requirements.txt
boto3
click
mlflow
numpy
pandas
Pillow
s3fs
sklearn
#tensorflow
json-tricks
